# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Notification Service API
  description: Event-driven notification system for processing and delivering messages through multiple channels (SMS, Email, Push).
  version: 1.0.0

servers:
  - url: http://localhost:8081
    description: Local development

paths:
  /health:
    get:
      summary: Health check (liveness)
      tags: [Health]
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /ready:
    get:
      summary: Readiness check
      tags: [Health]
      responses:
        "200":
          description: Service and dependencies are ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "503":
          description: Service is degraded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /metrics:
    get:
      summary: Real-time system metrics
      tags: [Observability]
      responses:
        "200":
          description: Current system metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetricsResponse"

  /api/v1/notifications:
    post:
      summary: Create a notification
      tags: [Notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationRequest"
      responses:
        "202":
          description: Notification accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      summary: List notifications with filtering and pagination
      tags: [Notifications]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, queued, processing, sent, failed, cancelled]
        - name: channel
          in: query
          schema:
            type: string
            enum: [sms, email, push]
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Paginated list of notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponse"

  /api/v1/notifications/batch:
    post:
      summary: Create a batch of notifications (up to 1000)
      tags: [Notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchCreateRequest"
      responses:
        "202":
          description: Batch accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchCreateResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/notifications/batch/{batchId}/status:
    get:
      summary: Get batch status summary
      tags: [Notifications]
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Batch status summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  batch_id:
                    type: string
                    format: uuid
                  total:
                    type: integer
                  statuses:
                    type: object
                    additionalProperties:
                      type: integer
        "404":
          description: Batch not found

  /api/v1/notifications/{id}:
    get:
      summary: Get notification by ID
      tags: [Notifications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Notification details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/v1/notifications/{id}/status:
    get:
      summary: Query notification status by ID
      tags: [Notifications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Notification status
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [pending, queued, processing, sent, failed, cancelled]
        "404":
          description: Not found

  /api/v1/notifications/{id}/cancel:
    patch:
      summary: Cancel a pending or queued notification
      tags: [Notifications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Notification cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "400":
          description: Cannot cancel (already sent/failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found

components:
  schemas:
    CreateNotificationRequest:
      type: object
      required: [channel, recipient, content]
      properties:
        channel:
          type: string
          enum: [sms, email, push]
        recipient:
          type: string
          description: Phone number, email address, or device token
        content:
          type: string
        priority:
          type: integer
          enum: [0, 1, 2]
          description: "0=high, 1=normal (default), 2=low"
          default: 1

    BatchCreateRequest:
      type: object
      required: [notifications]
      properties:
        notifications:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            $ref: "#/components/schemas/CreateNotificationRequest"

    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        batch_id:
          type: string
          format: uuid
        channel:
          type: string
          enum: [sms, email, push]
        recipient:
          type: string
        content:
          type: string
        priority:
          type: integer
        status:
          type: string
          enum: [pending, queued, processing, sent, failed, cancelled]
        attempt_count:
          type: integer
        max_attempts:
          type: integer
        last_error:
          type: string
        sent_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    BatchCreateResponse:
      type: object
      properties:
        batch_id:
          type: string
        total_created:
          type: integer
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"

    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        page:
          type: integer
        per_page:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer

    MetricsResponse:
      type: object
      properties:
        queue_depth:
          type: object
          additionalProperties:
            type: integer
        total_sent:
          type: integer
        total_failed:
          type: integer
        total_pending:
          type: integer
        success_rate:
          type: number
        failure_rate:
          type: number

    HealthResponse:
      type: object
      properties:
        status:
          type: string
        services:
          type: object
          additionalProperties:
            type: string

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details: {}